refactor: move table-structure-cleanup to a separate module
- we'll need it to mock mutation observer in jest
